CurvesObjectで使用オブジェクトを汎化
